I created two queues, TakeOff and. Landing LinkedList
Firstly, there is a loop that repeatedly prompts the user to input instructions.

Based on the input instruction, it calls the corresponding methods: takeoff(), landing(), Next(), or exits the program.

The takeoff() method adds the flight name to the queue for TakeOff. The landing() method also adds the flight name to the queue. The Next() method is to check if there are any planes waiting to land (Landing queue). If there is, it will delete the first aircraft from the queue and print the information. If there is no aircraft to land in the landing queue, the aircraft in the takeoff queue will be checked, and the first aircraft will be removed from the queue and printed as takeoff. If there are no planes in both queues, it will print that there are no planes in both queues.